+++ src/screens/auth/create-account/index.tsx
@@ -13,83 +13,93 @@ import {AuthHeader, BigButton, InputCustom} from '../../../components';
 import Header from '../../../components/customs/Headers';
 import {routes} from '../../../constants';
 import {NavigationService} from '../../../navigation';
-import {
-  checked,
-  toggleCheckbox,
-  credentials,
-  setCredentials,
-  inputErrors,
-  handleCreateAccount,
-} from '../../../utils/validate';
+// import {
+//   checked,
+//   toggleCheckbox,
+//   credentials,
+//   setCredentials,
+//   inputErrors,
+//   handleCreateAccount,
+// } from '../../../utils/validate';
 import useStyles from './styles';
+import {
+  comparePassword,
+  isValidEmail,
+  isValidPassword,
+  showToastError,
+} from '../../../utils';
+import {AuthActions} from '../../../redux';
+import {useAppDispatch} from '../../../hooks';
 
 const CreateAccount: FunctionComponent = () => {
   const styles = useStyles();
 
-  // const toggleCheckbox = () => setChecked(!checked);
-  // const dispatch = useAppDispatch();
-  // const [credentials, setCredentials] = React.useState<{
-  //   email: string;
-  //   password: string;
-  //   comfirmPassword: string;
-  // }>({
-  //   email: '',
-  //   password: '123456',
-  //   comfirmPassword: '123456',
-  // });
+  const [checked, setChecked] = React.useState<boolean>(false);
 
-  // const [inputErrors, setInputErrors] = React.useState<{
-  //   email: boolean;
-  //   password: boolean;
-  //   comfirmPassword: boolean;
-  // }>({
-  //   email: false,
-  //   password: false,
-  //   comfirmPassword: false,
-  // });
+  const toggleCheckbox = () => setChecked(!checked);
+  const dispatch = useAppDispatch();
+  const [credentials, setCredentials] = React.useState<{
+    email: string;
+    password: string;
+    comfirmPassword: string;
+  }>({
+    email: '',
+    password: '123456',
+    comfirmPassword: '123456',
+  });
 
-  // const handleCreateAccount = async () => {
-  //   const emailIsValid = isValidEmail(credentials.email);
-  //   const passwordIsValid = isValidPassword(credentials.password);
-  //   const comfirmPasswordIsValid = comparePassword(
-  //     credentials.password,
-  //     credentials.comfirmPassword,
-  //   );
-  //   setInputErrors({
-  //     email: !emailIsValid,
-  //     password: !passwordIsValid,
-  //     comfirmPassword: !comfirmPasswordIsValid,
-  //   });
-  //   if (
-  //     credentials.email.length === 0 ||
-  //     credentials.password.length === 0 ||
-  //     credentials.comfirmPassword.length === 0
-  //   ) {
-  //     showToastError('Please enter full information');
+  const [inputErrors, setInputErrors] = React.useState<{
+    email: boolean;
+    password: boolean;
+    comfirmPassword: boolean;
+  }>({
+    email: false,
+    password: false,
+    comfirmPassword: false,
+  });
 
-  //     return;
-  //   }
+  const handleCreateAccount = async () => {
+    const emailIsValid = isValidEmail(credentials.email);
+    const passwordIsValid = isValidPassword(credentials.password);
+    const comfirmPasswordIsValid = comparePassword(
+      credentials.password,
+      credentials.comfirmPassword,
+    );
+    setInputErrors({
+      email: !emailIsValid,
+      password: !passwordIsValid,
+      comfirmPassword: !comfirmPasswordIsValid,
+    });
+    if (
+      credentials.email.length === 0 ||
+      credentials.password.length === 0 ||
+      credentials.comfirmPassword.length === 0
+    ) {
+      showToastError('Please enter full information');
 
-  //   if (!emailIsValid) {
-  //     showToastError('Please enter the correct email format');
-  //     return;
-  //   }
-  //   if (!comfirmPasswordIsValid) {
-  //     showToastError('Password must be more than 6 characters');
-  //     return;
-  //   }
-  //   if (!passwordIsValid) {
-  //     showToastError('Confirm password does not match');
-  //     return;
-  //   }
+      return;
+    }
 
-  //   dispatch(
-  //     AuthActions.handleCreateAccount({
-  //       email: credentials.email,
-  //       password: credentials.password,
-  //     }),
-  //   );
-  // };
+    if (!emailIsValid) {
+      showToastError('Please enter the correct email format');
+      return;
+    }
+    if (!comfirmPasswordIsValid) {
+      showToastError('Password must be more than 6 characters');
+      return;
+    }
+    if (!passwordIsValid) {
+      showToastError('Confirm password does not match');
+      return;
+    }
+
+    dispatch(
+      AuthActions.handleCreateAccount({
+        email: credentials.email,
+        password: credentials.password,
+      }),
+    );
+  };
 
   return (
     <View style={styles.container}>
@@ -121,7 +131,6 @@ const CreateAccount: FunctionComponent = () => {
                 onChangeText={text =>
                   setCredentials({...credentials, email: text})
                 }
-                style={inputErrors.email ? styles.errorInput : null}
               />
               <Text style={styles.titleInput}>Password</Text>
               <InputCustom
